+++ crates/medea-reactive/src/collections/hash_set.rs
@@ -2,11 +2,13 @@
 
 use std::{collections::hash_set::Iter, hash::Hash, marker::PhantomData};
 
-use futures::{
-    future, future::LocalBoxFuture, stream::LocalBoxStream, FutureExt,
-};
+use futures::stream::LocalBoxStream;
 
-use crate::subscribers_store::{common, progressable, SubscribersStore};
+use crate::subscribers_store::{
+    common, progressable,
+    progressable::{RecheckableCounterFuture, RecheckableFutureExt},
+    SubscribersStore,
+};
 
 /// Reactive hash set based on [`HashSet`] with an ability to recognize when all
 /// updates was processed by subscribers.
@@ -114,40 +116,31 @@ impl<T> ProgressableHashSet<T>
 where
     T: Clone + 'static,
 {
-    /// Returns [`Future`] resolving when all push updates will be processed by
-    /// [`HashSet::on_insert()`] subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all push updates will be
+    /// processed by [`HashSet::on_insert()`] subscribers.
     #[inline]
     #[must_use]
-    pub fn when_insert_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_insert_processed(&self) -> RecheckableCounterFuture {
         self.on_insert_subs.when_all_processed()
     }
 
-    /// Returns [`Future`] resolving when all remove updates will be processed
-    /// by [`HashSet::on_remove()`] subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all remove updates will
+    /// be processed by [`HashSet::on_remove()`] subscribers.
     #[inline]
     #[must_use]
-    pub fn when_remove_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_remove_processed(&self) -> RecheckableCounterFuture {
         self.on_remove_subs.when_all_processed()
     }
 
-    /// Returns [`Future`] resolving when all insert and remove updates will be
-    /// processed by subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all insert and remove
+    /// updates will be processed by subscribers.
     #[inline]
     #[must_use]
-    pub fn when_all_processed(&self) -> LocalBoxFuture<'static, ()> {
-        Box::pin(
-            future::join(
-                self.when_remove_processed(),
-                self.when_insert_processed(),
-            )
-            .map(|(_, _)| ()),
-        )
+    pub fn when_all_processed(&self) -> impl RecheckableFutureExt<Output = ()> {
+        crate::join_all(vec![
+            self.when_remove_processed(),
+            self.when_insert_processed(),
+        ])
     }
 }