+++ crates/medea-reactive/src/collections/vec.rs
@@ -2,11 +2,13 @@
 
 use std::{marker::PhantomData, slice::Iter};
 
-use futures::{
-    future, future::LocalBoxFuture, stream::LocalBoxStream, FutureExt as _,
-};
+use futures::stream::LocalBoxStream;
 
-use crate::subscribers_store::{common, progressable, SubscribersStore};
+use crate::subscribers_store::{
+    common, progressable,
+    progressable::{RecheckableCounterFuture, RecheckableFutureExt},
+    SubscribersStore,
+};
 
 /// Reactive vector based on [`Vec`] with additional functionality of tracking
 /// progress made by its subscribers. Its [`Vec::on_push()`] and
@@ -97,40 +99,31 @@ impl<T> ProgressableVec<T>
 where
     T: Clone + 'static,
 {
-    /// Returns [`Future`] resolving when all push updates will be processed by
-    /// [`Vec::on_push()`] subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all push updates will be
+    /// processed by [`Vec::on_push()`] subscribers.
     #[inline]
     #[must_use]
-    pub fn when_push_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_push_processed(&self) -> RecheckableCounterFuture {
         self.on_push_subs.when_all_processed()
     }
 
-    /// Returns [`Future`] resolving when all remove updates will be processed
-    /// by [`Vec::on_remove()`] subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all remove updates will
+    /// be processed by [`Vec::on_remove()`] subscribers.
     #[inline]
     #[must_use]
-    pub fn when_remove_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_remove_processed(&self) -> RecheckableCounterFuture {
         self.on_remove_subs.when_all_processed()
     }
 
-    /// Returns [`Future`] resolving when all push and remove updates will be
-    /// processed by subscribers.
-    ///
-    /// [`Future`]: std::future::Future
+    /// Returns [`RecheckableFutureExt`] resolving when all push and remove
+    /// updates will be processed by subscribers.
     #[inline]
     #[must_use]
-    pub fn when_all_processed(&self) -> LocalBoxFuture<'static, ()> {
-        Box::pin(
-            future::join(
-                self.when_remove_processed(),
-                self.when_push_processed(),
-            )
-            .map(|(_, _)| ()),
-        )
+    pub fn when_all_processed(&self) -> impl RecheckableFutureExt<Output = ()> {
+        crate::join_all(vec![
+            self.when_remove_processed(),
+            self.when_push_processed(),
+        ])
     }
 }