+++ crates/medea-reactive/src/field/progressable_cell.rs
@@ -1,0 +1,102 @@
+//! Implementation of the progressable analogue of the [`Cell`].
+//!
+//! Subscription to changes works the same way as [`Progressable`],
+//! but working with underlying data of [`ProgressableCell`] is different.
+//!
+//! [`Cell`]: std::cell::Cell
+//! [`Progressable`]: crate::Progressable
+
+use std::cell::{Ref, RefCell};
+
+use futures::stream::LocalBoxStream;
+
+use crate::{
+    subscribers_store::{progressable, progressable::RecheckableCounterFuture},
+    Guarded, MutObservableFieldGuard,
+};
+
+use super::Progressable;
+
+/// Progressable analogue of [`Cell`].
+///
+/// Subscription to changes works the same way as [`Progressable`], but working
+/// with underlying data of [`ProgressableCell`] is different.
+///
+/// [`Cell`]: std::cell::Cell
+#[derive(Debug)]
+pub struct ProgressableCell<D>(RefCell<Progressable<D>>);
+
+impl<D> ProgressableCell<D>
+where
+    D: 'static,
+{
+    /// Returns new [`ProgressableCell`] with subscribable mutations.
+    #[inline]
+    pub fn new(data: D) -> Self {
+        Self(RefCell::new(Progressable::new(data)))
+    }
+
+    /// Returns immutable reference to an underlying data.
+    #[inline]
+    pub fn borrow(&self) -> Ref<'_, D> {
+        let reference = self.0.borrow();
+        Ref::map(reference, |observable| &**observable)
+    }
+}
+
+impl<D> ProgressableCell<D>
+where
+    D: Clone + 'static,
+{
+    /// Returns copy of an underlying data.
+    #[inline]
+    pub fn get(&self) -> D {
+        self.0.borrow().data.clone()
+    }
+
+    /// Returns [`Stream`] into which underlying data updates will be emitted.
+    ///
+    /// [`Stream`]: futures::Stream
+    #[inline]
+    pub fn subscribe(&self) -> LocalBoxStream<'static, Guarded<D>> {
+        self.0.borrow().subscribe()
+    }
+
+    /// Returns [`RecheckableFutureExt`] which will be resolved when all data
+    /// updates will be processed by subscribers.
+    ///
+    /// [`RecheckableFutureExt`]: crate::RecheckableFutureExt
+    #[inline]
+    pub fn when_all_processed(&self) -> RecheckableCounterFuture {
+        self.0.borrow().when_all_processed()
+    }
+}
+
+impl<D> ProgressableCell<D>
+where
+    D: Clone + PartialEq + 'static,
+{
+    /// Sets the `new_data` value as an underlying data.
+    #[inline]
+    pub fn set(&self, new_data: D) {
+        *self.0.borrow_mut().borrow_mut() = new_data;
+    }
+
+    /// Replaces the contained underlying data with the given `new_data` value,
+    /// and returns the old one.
+    #[inline]
+    pub fn replace(&self, mut new_data: D) -> D {
+        std::mem::swap(&mut *self.0.borrow_mut().borrow_mut(), &mut new_data);
+        new_data
+    }
+
+    /// Updates an underlying data using the provided function, which will
+    /// accept a mutable reference to an underlying data.
+    #[inline]
+    pub fn mutate<F>(&self, f: F)
+    where
+        F: FnOnce(MutObservableFieldGuard<'_, D, progressable::SubStore<D>>),
+    {
+        (f)(self.0.borrow_mut().borrow_mut());
+    }
+}