+++ crates/medea-reactive/CHANGELOG.md
@@ -23,10 +23,13 @@ All user visible changes to this project will be documented in this file. This p
     - `ProgressableVec`;
     - `ProgressableHashMap`;
     - `ProgressableHashSet`.
+- `ProgressableCell` container ([#170]);
+- `RecheckableFutureExt` `Future` extension ([#170]).
 
 [#81]: /../../pull/81
 [#103]: /../../pull/103
 [#159]: /../../pull/159
+[#170]: /../../pull/170